Extract _schema.py from _legacy.py
